requirements.txt
cffi==1.15.1
click==8.1.3
cryptography==39.0.2
Flask==2.2.3
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-MySQL==1.5.2
greenlet==2.0.2
gunicorn==20.1.0
importlib-metadata==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
marshmallow==3.19.0
packaging==23.0
protobuf==3.20.3
psycopg2-binary==2.9.5
pycparser==2.21
PyMySQL==1.0.2
python-dotenv==1.0.0
six==1.16.0
typing-extensions==4.5.0
Werkzeug==2.2.3
zipp==3.15.0